chore(*) - Cleanup code for aws - Removed vagrant specific settings for amazon-ebs - Corrected template.json to proper amazon name - removed cleanup.sh from aws - removed virtualbox.sh until can remove veewee specific configurations

+++ centos64_puppet/scripts/cleanup.sh
@@ -3,4 +3,7 @@ yum -y clean all
 rm -rf /etc/yum.repos.d/{puppetlabs,epel}.repo
 rm -rf VBoxGuestAdditions_*.iso
 sed -i /HWADDR/d /etc/sysconfig/network-scripts/ifcfg-eth0
-rm /etc/udev/rules.d/70-persistent-net.rules
+
+if [[ -f /etc/udev/rules.d/70-persistent-net.rules ]] then 
+  rm /etc/udev/rules.d/70-persistent-net.rules
+fi
